voice-service/requirements.txt
# Voice Service

fastapi>=0.109.0
uvicorn>=0.27.0
httpx>=0.26.0
pydantic>=2.5.0
python-multipart>=0.0.6
websockets>=12.0
python-dotenv>=1.0.0
